+++ osfmk/corecrypto/cchmac/src/cchmac_update.c
@@ -11,6 +11,6 @@
 #include <corecrypto/cchmac.h>
 
 void cchmac_update(const struct ccdigest_info *di, cchmac_ctx_t hc,
-                   unsigned long data_len, const void *data) {
+                   size_t data_len, const void *data) {
     ccdigest_update(di, cchmac_digest_ctx(di, hc), data_len, data);
 }
